requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.68.2
furl==2.1.3
h11==0.13.0
idna==3.3
install==1.3.5
orderedmultidict==1.0.1
outcome==1.2.0
pydantic==1.9.2
PySocks==1.7.1
PyYAML==6.0
requests==2.28.1
selenium==4.4.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
starlette==0.14.2
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.15.0
wsproto==1.2.0
